+++ features/Context/FixturesContext.php
@@ -13,7 +13,6 @@ use Oro\Bundle\UserBundle\Entity\User;
 use Pim\Bundle\CatalogBundle\Entity\AssociationType;
 use Pim\Bundle\CatalogBundle\Entity\AttributeGroup;
 use Pim\Bundle\CatalogBundle\Entity\AttributeOption;
-use Pim\Bundle\CatalogBundle\Entity\AttributeRequirement;
 use Pim\Bundle\CatalogBundle\Entity\GroupType;
 use Pim\Bundle\CatalogBundle\Entity\Channel;
 use Pim\Bundle\CatalogBundle\Entity\Family;
@@ -676,32 +675,6 @@ class FixturesContext extends RawMinkContext
         }
 
         $jobInstance->setRawConfiguration($configuration);
-        $this->flush();
-    }
-
-    /**
-     * @param TableNode $table
-     *
-     * @Given /^the following attribute requirements:$/
-     */
-    public function theFollowingAttributeRequirements(TableNode $table)
-    {
-        foreach ($table->getHash() as $data) {
-            $requirement = new AttributeRequirement();
-
-            $attribute = $this->getAttribute($data['attribute']);
-            $channel   = $this->getChannel($data['scope']);
-            $family    = $this->getFamily($data['family']);
-
-            $requirement->setAttribute($attribute);
-            $requirement->setChannel($channel);
-            $requirement->setFamily($family);
-
-            $requirement->setRequired($data['required'] === 'yes');
-
-            $this->persist($requirement, false);
-        }
-
         $this->flush();
     }
 
@@ -1410,7 +1383,12 @@ class FixturesContext extends RawMinkContext
         $categories = $this->loadFixture('categories', $data);
 
         foreach ($categories as $category) {
-            $this->persist($category);
+            $this->persist($category, true);
+            foreach ($category->getProducts() as $product) {
+                $product->addCategory($category);
+                $this->flush($product);
+            }
+
         }
 
         return reset($categories);
@@ -1483,15 +1461,16 @@ class FixturesContext extends RawMinkContext
             $group->addAttribute($attribute);
         }
 
+        $this->persist($group);
+        $this->flush($group);
+
         foreach ($products as $sku) {
             if (!empty($sku)) {
                 $product = $this->getProduct($sku);
-                $group->addProduct($product);
                 $product->addGroup($group);
+                $this->flush($product);
             }
         }
-
-        $this->persist($group);
     }
 
     /**
@@ -1673,16 +1652,6 @@ class FixturesContext extends RawMinkContext
     }
 
     /**
-     * Flush
-     */
-    private function flush()
-    {
-        foreach ($this->getSmartRegistry()->getManagers() as $manager) {
-            $manager->flush();
-        }
-    }
-
-    /**
      * @return \Doctrine\ORM\EntityManager
      */
     private function getEntityManager()
@@ -1798,6 +1767,31 @@ class FixturesContext extends RawMinkContext
         $manager->remove($object);
 
         if ($flush) {
+            $manager->flush();
+        }
+    }
+
+    /**
+     * @param object $object
+     *
+     * @return null
+     */
+    private function flush($object = null)
+    {
+        if (!$object) {
+            return $this->flushAll();
+        }
+
+        $manager = $this->getSmartRegistry()->getManagerForClass(get_class($object));
+        $manager->flush($object);
+    }
+
+    /**
+     * Flush all managers
+     */
+    private function flushAll()
+    {
+        foreach ($this->getSmartRegistry()->getManagers() as $manager) {
             $manager->flush();
         }
     }