+++ spec/Pim/Bundle/CatalogBundle/MongoDB/Normalizer/ProductNormalizerSpec.php
@@ -47,7 +47,6 @@ class ProductNormalizerSpec extends ObjectBehavior
 
         $this->normalize($product, 'mongodb_json', [])->shouldReturn([
             ProductNormalizer::FAMILY_FIELD => 'family normalization',
-            ProductNormalizer::GROUPS_FIELD => [],
             ProductNormalizer::COMPLETENESSES_FIELD => array('completenessCode' => 'completeness normalization'),
             ProductNormalizer::ENABLED_FIELD => 1
         ]);